Added framework for player tournament commands.
