+++ src/tournament_commands/admin_commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod add_deck;
+pub mod admin;
 pub mod confirm_result;
 pub mod create_match;
 pub mod cut;
