+++ src/tournament_commands/player_commands/add_deck.rs
@@ -1,0 +1,10 @@
+use serenity::framework::standard::{macros::command, Args, CommandResult};
+use serenity::model::prelude::*;
+use serenity::prelude::*;
+
+#[command("add-deck")]
+#[description("Submits a deck.")]
+async fn add_deck(ctx: &Context, msg: &Message, _: Args) -> CommandResult {
+    todo!()
+}
+
