+++ src/tournament_commands/player_commands/match_result.rs
@@ -1,0 +1,11 @@
+use serenity::framework::standard::{macros::command, Args, CommandResult};
+use serenity::model::prelude::*;
+use serenity::prelude::*;
+
+#[command("match-result")]
+#[only_in(guild)]
+#[description("Submit the result of a match.")]
+async fn match_result(ctx: &Context, msg: &Message, _: Args) -> CommandResult {
+    todo!()
+}
+
